requirements.txt
asgiref==3.8.1
dj-database-url==2.3.0
django==4.2.20
django-heroku==0.3.1
gunicorn==23.0.0
packaging==24.2
pillow==11.2.0
psycopg-binary==3.2.6
psycopg2==2.9.10
psycopg2-binary==2.9.10
sqlparse==0.5.1
typing_extensions==4.12.2
whitenoise==6.8.2